+++ frontend/src/app/pages/exercises/exercise/shared/simulated-region-overview/tabs/behavior-tab/utils/treatment-progress-to-german-name.pipe.ts
@@ -5,6 +5,7 @@ const treatmentProgressToGermanNameDictionary: {
     [Key in TreatmentProgress]: string;
 } = {
     counted: 'Vorsichten',
+    noTreatment: 'Keine Behandlung',
     secured: 'Erstversorgung sichergestellt',
     triaged: 'Behandeln, Personal fehlt',
     unknown: 'Erkunden',