+++ shared/src/models/radiogram/status/radiogram-status.ts
@@ -1,0 +1,3 @@
+export class RadiogramStatus {
+    readonly type!: `${string}RadiogramStatus`;
+}