+++ shared/src/simulation/activities/exercise-simulation-activity.ts
@@ -3,11 +3,15 @@ import { delayEventActivity } from './delay-event';
 import { SimulationActivityState } from './simulation-activity';
 import { reassignTreatmentsActivity } from './reassign-treatments';
 import { unloadVehicleActivity } from './unload-vehicle';
+import { recurringEventActivity } from './recurring-event';
+import { generateReportActivity } from './generate-report';
 
 export const simulationActivities = {
     reassignTreatmentsActivity,
     unloadVehicleActivity,
     delayEventActivity,
+    recurringEventActivity,
+    generateReportActivity,
 };
 
 export type ExerciseSimulationActivity =