+++ shared/src/simulation/behaviors/treat-patients.ts
@@ -6,7 +6,17 @@ import {
     Min,
     ValidateNested,
 } from 'class-validator';
-import { getCreate } from '../../models/utils';
+import { groupBy } from 'lodash-es';
+import { Patient } from '../../models';
+import type {
+    PatientCountRadiogram,
+    TreatmentStatusRadiogram,
+} from '../../models/radiogram';
+import { getCreate, isInSpecificSimulatedRegion } from '../../models/utils';
+import type { SimulatedRegion } from '../../models';
+import type { ExerciseState } from '../../state';
+import { getActivityById } from '../../store/action-reducers/utils';
+import type { Mutable } from '../../utils';
 import { uuid, UUID, uuidValidationOptions } from '../../utils';
 import { IsLiteralUnion, IsValue } from '../../utils/validators';
 import { DelayEventActivityState } from '../activities';
@@ -104,7 +114,7 @@ export class TreatPatientsBehaviorState implements SimulationBehaviorState {
     public readonly treatmentActivityId: UUID | null = null;
 
     @IsLiteralUnion(treatmentProgressAllowedValues)
-    public readonly treatmentProgress: TreatmentProgress = 'unknown';
+    public readonly treatmentProgress: TreatmentProgress = 'noTreatment';
 
     static readonly create = getCreate(this);
 }
@@ -115,6 +125,15 @@ export const treatPatientsBehavior: SimulationBehavior<TreatPatientsBehaviorStat
         handleEvent(draftState, simulatedRegion, behaviorState, event) {
             switch (event.type) {
                 case 'tickEvent':
+                    if (behaviorState.treatmentProgress === 'noTreatment') {
+                        startNewTreatmentReassignment(
+                            draftState,
+                            simulatedRegion,
+                            behaviorState
+                        );
+                        break;
+                    }
+
                     if (
                         behaviorState.delayActivityId === null &&
                         (behaviorState.treatmentActivityId === null ||
@@ -157,37 +176,122 @@ export const treatPatientsBehavior: SimulationBehavior<TreatPatientsBehaviorStat
                         behaviorState.delayActivityId = null;
                     }
 
-                    if (
-                        behaviorState.treatmentActivityId &&
-                        simulatedRegion.activities[
-                            behaviorState.treatmentActivityId
-                        ]
-                    ) {
-                        terminateActivity(
-                            draftState,
-                            simulatedRegion,
-                            behaviorState.treatmentActivityId
-                        );
-                        behaviorState.treatmentActivityId = null;
-                    }
-
-                    const id = nextUUID(draftState);
-                    addActivity(
+                    startNewTreatmentReassignment(
+                        draftState,
                         simulatedRegion,
-                        ReassignTreatmentsActivityState.create(
-                            id,
-                            behaviorState.treatmentProgress,
-                            behaviorState.intervals.countingTimePerPatient
-                        )
+                        behaviorState
                     );
-                    behaviorState.treatmentActivityId = id;
+
                     break;
                 }
                 case 'treatmentProgressChangedEvent':
                     behaviorState.treatmentProgress = event.newProgress;
                     break;
+                case 'collectInformationEvent': {
+                    const collectInformationEvent = event;
+
+                    const radiogram = getActivityById(
+                        draftState,
+                        simulatedRegion.id,
+                        event.generateReportActivityId,
+                        'generateReportActivity'
+                    ).radiogram;
+
+                    switch (collectInformationEvent.informationType) {
+                        // This behavior answerers this query because the treating personnel has the knowledge of how many patients are in a given category
+                        case 'patientCount': {
+                            if (behaviorState.treatmentProgress === 'unknown') {
+                                return;
+                            }
+
+                            const patientCountRadiogram =
+                                radiogram as Mutable<PatientCountRadiogram>;
+
+                            const patientCount =
+                                patientCountRadiogram.patientCount;
+                            const patients = Object.values(
+                                draftState.patients
+                            ).filter((patient) =>
+                                isInSpecificSimulatedRegion(
+                                    patient,
+                                    simulatedRegion.id
+                                )
+                            );
+                            const groupedPatients = groupBy(
+                                patients,
+                                (patient) =>
+                                    Patient.getVisibleStatus(
+                                        patient,
+                                        draftState.configuration
+                                            .pretriageEnabled,
+                                        draftState.configuration
+                                            .bluePatientsEnabled
+                                    )
+                            );
+                            patientCount.black =
+                                groupedPatients['black']?.length ?? 0;
+                            patientCount.white =
+                                groupedPatients['white']?.length ?? 0;
+                            patientCount.red =
+                                groupedPatients['red']?.length ?? 0;
+                            patientCount.yellow =
+                                groupedPatients['yellow']?.length ?? 0;
+                            patientCount.green =
+                                groupedPatients['green']?.length ?? 0;
+                            patientCount.blue =
+                                groupedPatients['blue']?.length ?? 0;
+
+                            patientCountRadiogram.informationAvailable = true;
+                            break;
+                        }
+                        case 'treatmentStatus': {
+                            const treatmentStatusRadiogram =
+                                radiogram as Mutable<TreatmentStatusRadiogram>;
+
+                            treatmentStatusRadiogram.treatmentStatus =
+                                behaviorState.treatmentProgress;
+
+                            treatmentStatusRadiogram.informationAvailable =
+                                true;
+                            break;
+                        }
+                        default:
+                        // Ignore event, since this behavior can't answer this query
+                    }
+
+                    break;
+                }
                 default:
                 // Ignore event
             }
         },
     };
+
+function startNewTreatmentReassignment(
+    draftState: Mutable<ExerciseState>,
+    simulatedRegion: Mutable<SimulatedRegion>,
+    behaviorState: Mutable<TreatPatientsBehaviorState>
+) {
+    if (
+        behaviorState.treatmentActivityId &&
+        simulatedRegion.activities[behaviorState.treatmentActivityId]
+    ) {
+        terminateActivity(
+            draftState,
+            simulatedRegion,
+            behaviorState.treatmentActivityId
+        );
+        behaviorState.treatmentActivityId = null;
+    }
+
+    const id = nextUUID(draftState);
+    addActivity(
+        simulatedRegion,
+        ReassignTreatmentsActivityState.create(
+            id,
+            behaviorState.treatmentProgress,
+            behaviorState.intervals.countingTimePerPatient
+        )
+    );
+    behaviorState.treatmentActivityId = id;
+}